+++ connect/amazon/iamidentitycenter/fake/src/test/kotlin/org/http4k/connect/amazon/iamidentitycenter/FakeIAMIdentityCenterTest.kt
@@ -2,6 +2,7 @@ package org.http4k.connect.amazon.iamidentitycenter
 
 import org.http4k.connect.amazon.FakeAwsContract
 import org.http4k.core.HttpHandler
+import org.http4k.routing.ReverseProxyHostMatcher.Companion.Contains
 import org.http4k.routing.reverseProxy
 import kotlin.io.path.Path
 
@@ -14,6 +15,7 @@ class FakeIAMIdentityCenterTest : IAMIdentityCenterContract, FakeAwsContract {
 
     override val http: HttpHandler = reverseProxy(
         "sso" to FakeSSO(),
-        "oidc" to FakeOIDC()
+        "oidc" to FakeOIDC(),
+        matcher = Contains
     )
 }